pyproject.toml
[project]
name = "agent-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.2.0",
    "funky-sdk==0.2.8",
    "imageio[ffmpeg]>=2.37.3",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "langsmith>=0.7.30",
    "numpy>=2.4.4",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
